requirements.txt
azure-functions==1.24.0 ; python_version >= "3.12"
markupsafe==3.0.3 ; python_version >= "3.12"
werkzeug==3.1.6 ; python_version >= "3.12"
